requirements.txt
fastapi
chromadb 
uvicorn
python-dotenv
langchain
langchain-openai
pydantic
openai